Added FAB menu lib for options, moved generate qr from main view into generate qr view, stubbed out scan qr view
