+++ qrCode/app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".views.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -22,6 +22,10 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+        <activity android:name=".views.GenerateQrCodeActivity"
+                  android:label="Generate QR Code" />
+        <activity android:name=".views.ScanQrCodeActivity"
+                  android:label="Scan QR Code" />
     </application>
 
 </manifest>
