--- qrCode/app/src/main/java/com/superformula/qrcode/MainActivity.kt
@@ -1,62 +1,0 @@
-package com.superformula.qrcode
-
-import android.arch.lifecycle.ViewModelProviders
-import android.os.Bundle
-import android.support.design.widget.Snackbar
-import android.support.v7.app.AppCompatActivity
-import android.view.Menu
-import android.view.MenuItem
-import android.view.View
-import com.superformula.qrcode.models.Seed
-import com.superformula.qrcode.viewModels.QrViewModel
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.content_main.*
-import org.joda.time.DateTime
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var viewModel: QrViewModel
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        setSupportActionBar(toolbar)
-
-        viewModel = ViewModelProviders.of(this).get(QrViewModel::class.java)
-
-        viewModel.loadSeed().subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe{ seed -> showQrCode(seed) }
-
-        fab.setOnClickListener { view ->
-            Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                    .setAction("Action", null).show()
-        }
-    }
-
-    private fun showQrCode(seed: Seed) {
-        qrImage.setImageBitmap(viewModel.generateQrCode(seed))
-        progressSpinner.visibility = View.GONE
-        qrImage.visibility = View.VISIBLE
-        val expireDate = DateTime.parse((seed.expiresAt))
-
-        expireText.text = String.format("Expires: %s", expireDate.toLocalDateTime().toString("MM/dd/yyyy HH:mm:ss"))
-    }
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        menuInflater.inflate(R.menu.menu_main, menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        return when (item.itemId) {
-            R.id.action_settings -> true
-            else -> super.onOptionsItemSelected(item)
-        }
-    }
-}
