+++ qrCode/app/src/main/java/com/superformula/qrcode/views/MainActivity.kt
@@ -1,0 +1,45 @@
+package com.superformula.qrcode.views
+
+import android.content.Intent
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import android.view.Menu
+import android.view.MenuItem
+import com.superformula.qrcode.R
+import kotlinx.android.synthetic.main.activity_main.*
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        setSupportActionBar(toolbar)
+
+        action_menu.addOnMenuItemClickListener({ _, _, floatingActionButton ->
+            if (floatingActionButton.id == R.id.scan_qr) {
+                val intent = Intent(this, ScanQrCodeActivity::class.java)
+                startActivity(intent)
+            }
+            if (floatingActionButton.id == R.id.generate_qr) {
+                val intent = Intent(this, GenerateQrCodeActivity::class.java)
+                startActivity(intent)
+            }
+        })
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        menuInflater.inflate(R.menu.menu_main, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        return when (item.itemId) {
+            R.id.action_settings -> true
+            else -> super.onOptionsItemSelected(item)
+        }
+    }
+}
