+++ qrCode/app/src/main/res/drawable/ic_scanner_white_24dp.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+        android:width="24dp"
+        android:height="24dp"
+        android:viewportWidth="24.0"
+        android:viewportHeight="24.0">
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M19.8,10.7L4.2,5l-0.7,1.9L17.6,12L5,12c-1.1,0 -2,0.9 -2,2v4c0,1.1 0.9,2 2,2h14c1.1,0 2,-0.9 2,-2v-5.5c0,-0.8 -0.5,-1.6 -1.2,-1.8zM7,17L5,17v-2h2v2zM19,17L9,17v-2h10v2z"/>
+</vector>
