+++ qrCode/app/src/main/res/menu/fab_menu_options.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!-- First button is the initial Fab of the menu -->
+    <!-- Title here can be empty if you don't to add any action in the main button -->
+    <item
+        android:id="@+id/main_option"
+        android:icon="@drawable/ic_add_white_24dp" />
+
+    <!-- Options buttons of the Fab menu -->
+    <item
+        android:id="@+id/generate"
+        android:icon="@drawable/ic_code_white_24dp"
+        android:title="QR Code" />
+
+    <item
+        android:id="@+id/scan"
+        android:icon="@drawable/ic_scanner_white_24dp"
+        android:title="Scan" />
+
+</menu>
